requirements.txt
tiktoken==0.3.3
deeplake==3.6.1
langchain==0.0.194
jupyter
notebook
py-solc-x==1.1.1
python-dotenv==1.0.0